profileservice/requirements/base.txt
django==1.11.17
djangorestframework==3.7.3
django-filter==1.1.0
Markdown==2.6.9
pytz==2017.3
